+++ src/utils/SimpleConditionsEnforcer.sol
@@ -20,17 +20,14 @@ contract SimpleConditionsEnforcer {
     /// @dev Compares the current block timestamp with a given timestamp.
     /// @param ct The type of comparison to be made.
     /// @param timestamp The timestamp to compare with.
-    function currentBlockTimestamp(
-        ComparisonType ct,
-        uint timestamp
-    ) external view {
+    function currentBlockTimestamp(ComparisonType ct, uint256 timestamp) external view {
         compare(block.timestamp, ct, timestamp);
     }
 
     /// @dev Compares the current block number with a given number.
     /// @param ct The type of comparison to be made.
     /// @param number The number to compare with.
-    function currentBlockNumber(ComparisonType ct, uint number) external view {
+    function currentBlockNumber(ComparisonType ct, uint256 number) external view {
         compare(block.number, ct, number);
     }
 
@@ -38,11 +35,7 @@ contract SimpleConditionsEnforcer {
     /// @param value1 The first value to compare.
     /// @param ct The type of comparison to be made.
     /// @param value2 The second value to compare.
-    function compare(
-        uint value1,
-        ComparisonType ct,
-        uint value2
-    ) internal pure {
+    function compare(uint256 value1, ComparisonType ct, uint256 value2) internal pure {
         if (ct == ComparisonType.EQ) {
             if (value1 != value2) revert ConditionNotMet();
         } else if (ct == ComparisonType.GT) {
